+++ TouchGFX/gui/include/gui/screen1_screen/Screen1View.hpp
@@ -12,7 +12,7 @@ public:
     virtual void setupScreen();
     virtual void tearDownScreen();
 
-    void UpdateGraph(float* value);
+    void UpdateGraph(uint8_t* value);
 protected:
 };
